clatd support for networkmanager environments
Systems that use network manager (typically clients, which is where
clatd will be deployed) have some unique issues with clatd, because
nm controls the whole stack in ways that break clatd. Generally this
means that nm disables the forwarding flags on the interfaces, which
causes traffic to not operate as expected.

This adds support to manually reset the plat dev forwarding flag
in the situation that nm or other tools have disabled it.